last burst of mini refactors
