+++ Java/src/main/java/com/gildedrose/categories/SulfurasCategory.java
@@ -1,4 +1,6 @@
-package com.gildedrose;
+package com.gildedrose.categories;
+
+import com.gildedrose.Item;
 
 public class SulfurasCategory implements ItemCategory {
